FEAT: Testando o fetch da api rodando
