+++ frontend/lib/main.dart
@@ -1,24 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:vida_leve/telas/autenticacao_tela.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'firebase_options.dart';
+import 'package:vida_leve/servicos/api_service.dart';
 
-void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp(
-    options: DefaultFirebaseOptions.currentPlatform,
-  );
+void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Autenticacao(),
+      home: Autenticacao(), // Tela principal
     );
   }
 }
