+++ frontend/pubspec.yaml
@@ -13,6 +13,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  http: ^0.13.5                              
 
   cupertino_icons: ^1.0.6
   firebase_core: ^2.13.1
